Introduce `Persistable` trait
This trait is meant to be impl'd by any pub payjoin type that is meant to be
persisted by the application. The trait has an associated `Key` type
which uniquely identifies the type. The trait imposes no specific
serialization format. The format depends on what the applications
intends to do with the data.

Issue: https://github.com/payjoin/rust-payjoin/issues/336