Return Install build-dependencies to the ubuntu-latest workflow
I moved this code incorrectly in #989 to a general binding build script but
others not using a debian based system immediately ran into problems running
this apt based package manager on their machine.

It was removed in #1000 from the binding script but This probably should
just be returned to the workflow it was originally.